+++ test/reporters/console.js
@@ -39,11 +39,10 @@ buster.testCase('console - _success _failure _end', {
     assert.equals(reporter.successes[0].test.command, 'successcommand1');
     assert.equals(reporter.successes[0].result.exitCode, 0);
 
-
     // should write a red dot and add test result to failures list
     this.mockStdout.expects('write').withExactArgs('.'.red);
 
-    reporter.emit('failure', ['someerror1' ], { description: 'failuredesc1', file: 'failure1.yml', command: 'failurecommand1' }, { exitcode: 10, output: 'failureoutput1' });
+    reporter.emit('failure', ['someerror1' ], { description: 'failuredesc1', file: 'failure1.yml', dir: 'failuredir1', command: 'failurecommand1' }, { exitcode: 10, output: 'failureoutput1' });
 
     assert.equals(reporter.failures.length, 1);
     assert.equals(reporter.failures[0].errors, ['someerror1']);
@@ -53,10 +52,11 @@ buster.testCase('console - _success _failure _end', {
     // should log summary in red text when there is failure, and debug output and exitcode
     this.mockConsole.expects('log').withExactArgs('');
     this.mockConsole.expects('error').withExactArgs('2 tests, 1 success, 1 failure'.red);
-    this.mockConsole.expects('log').withExactArgs('\n---\nfailuredesc1\n' + 'failure1.yml'.cyan + '> failurecommand1');
+    this.mockConsole.expects('error').withExactArgs('\n----------------\nfailuredesc1\n' + 'failure1.yml'.cyan + '\nfailurecommand1');
     this.mockConsole.expects('error').withExactArgs('someerror1'.red);
+    this.mockConsole.expects('error').withExactArgs('exec dir: %s\n'.grey, 'failuredir1');
+    this.mockConsole.expects('error').withExactArgs('exitcode: %d\n'.grey, 10);
     this.mockConsole.expects('error').withExactArgs('output:\n%s'.grey, 'failureoutput1');
-    this.mockConsole.expects('error').withExactArgs('exitcode: %d'.grey, 10);
 
     reporter.emit('end', true);
 
@@ -73,7 +73,7 @@ buster.testCase('console - _success _failure _end', {
     // emit another failure event, but this time without test description
     this.mockStdout.expects('write').withExactArgs('.'.red);
 
-    reporter.emit('failure', ['someerror2' ], { file: 'failure2.yml', command: 'failurecommand2' }, { exitcode: 10 });
+    reporter.emit('failure', ['someerror2' ], { file: 'failure2.yml', dir: 'failuredir2', command: 'failurecommand2' }, { exitcode: 10 });
 
     assert.equals(reporter.failures.length, 2);
     assert.equals(reporter.failures[1].errors, ['someerror2']);
@@ -83,13 +83,12 @@ buster.testCase('console - _success _failure _end', {
     // emit another end event, should not display description
     this.mockConsole.expects('log').withExactArgs('');
     this.mockConsole.expects('error').withExactArgs('4 tests, 2 successes, 2 failures'.red);
-    this.mockConsole.expects('log').withExactArgs('\n---\nfailuredesc1\n' + 'failure1.yml'.cyan + '> failurecommand1');
+    this.mockConsole.expects('error').withExactArgs('\n----------------\nfailuredesc1\n' + 'failure1.yml'.cyan + '\nfailurecommand1');
     this.mockConsole.expects('error').withExactArgs('someerror1'.red);
-    this.mockConsole.expects('log').withExactArgs('\n---\n' + 'failure2.yml'.cyan + '> failurecommand2');
+    this.mockConsole.expects('error').withExactArgs('\n----------------\n' + 'failure2.yml'.cyan + '\nfailurecommand2');
     this.mockConsole.expects('error').withExactArgs('someerror2'.red);
 
     reporter.emit('end', false);
-
   }
 });