+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add test execution directory to debug info
 * Add stdout and stderr test fields
 * Add environment variables to test commands parameters
+* Add test fixtures support
 
 ### 0.0.5
 * Fix sample test file on README.